Removing the hand written SQL statement for getObjects
